Update JPAUnitTestCase.java
There are 2 ways to make this test pass:

1 - [recomended] add a getter/setter to the entity

2 - enable ''extended enhancement" on the build plugin

By default, bytecode manipulation is restricted to classes that are entities. That is to keep the number of modified classes to the minimum. This requires that all the fields for the persistent attributes to be manipulated only in the entity itself (which is the right way to program under the OO paradigm). 

There are cases where the code can't be modified and for those scenarios there is 'extended enhancement' that convert the field access to the right method call. It's used throughout the test suite to reduce the boilerplate, so it's tested and works well, but it's not recommended because of the reasons mentioned.

I hope this helps you solve your problem and I'm happy to help with anything related to bytecode enhacement!